% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Draw the name with frequency over time}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{prenoms}

\item{the_sex}{sex of people}
}
\value{
a graph
}
\description{
Draw the name with frequency over time
}
\examples{
\dontrun{
draw_a_name()

}


}
