% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Use dygraph to draw the dynamic graph of the frequency of the}
\usage{
draw_names(names)
}
\arguments{
\item{names}{prenoms}
}
\value{
a graph
}
\description{
Use dygraph to draw the dynamic graph of the frequency of the
}
\examples{
\dontrun{
draw_names()
}
}
