% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save your files as csv}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{dataset to save}

\item{filename}{your file names}

\item{row.names}{name of rows}

\item{...}{any argument}
}
\value{
a list
}
\description{
Save your files as csv
}
\examples{
\dontrun{
save_as_csv()
}
}
