% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW06_q2_q3.R
\name{side_length}
\alias{side_length}
\title{side_length Function}
\usage{
side_length(a, b)
}
\arguments{
\item{a:}{length of one side of a triangle.No default.}

\item{b:}{length of one side of a triangle.No default.}
}
\description{
This function allows you to calculates the length of the third side according to the Pythagorean theorem, given the lengths of two sides of the triangle,
}
\examples{
side_length(3,4)

}
\keyword{side_length}
