% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW06_q2_q3.R
\name{trim_mean}
\alias{trim_mean}
\title{trim_mean  Function}
\usage{
trim_mean(x, s, l)
}
\arguments{
\item{x:}{a numeric vector.}

\item{s:}{s smallest values.}

\item{l:}{l largest values.}
}
\description{
This function allows you to calculates the mean of a numeric vector x, ignoring the s smallest and l largest values.
}
\examples{
trim_mean(x,1,2)

}
\keyword{trim_mean}
